requirements.txt
Flask
python-amazon-paapi